minor improvements

+++ core/src/de/dakror/vloxlands/util/math/Interval.java
@@ -12,17 +12,17 @@ public class Interval implements Comparable<Interval>, Savable
 {
 	int start, length;
 	byte type;
-
+	
 	public Interval()
 	{}
-
+	
 	public Interval(byte type, int start, int length)
 	{
 		this.type = type;
 		this.start = start;
 		this.length = length;
 	}
-
+	
 	public Interval set(byte type, int start, int length)
 	{
 		this.type = type;
@@ -30,63 +30,63 @@ public class Interval implements Comparable<Interval>, Savable
 		this.length = length;
 		return this;
 	}
-	
+
 	public byte getType()
 	{
 		return type;
 	}
-
+	
 	public void setType(byte type)
 	{
 		this.type = type;
 	}
-	
+
 	public void shrink()
 	{
 		length--;
 	}
-
+	
 	public void extend()
 	{
 		length++;
 	}
-
+	
 	public int getStart()
 	{
 		return start;
 	}
-
+	
 	public void setStart(int start)
 	{
 		this.start = start;
 	}
-
+	
 	public int getLength()
 	{
 		return length;
 	}
-
+	
 	public void setLength(int length)
 	{
 		this.length = length;
 	}
-
+	
 	@Override
 	public int compareTo(Interval o)
 	{
 		return Integer.compare(start, o.start);
 	}
-	
+
 	@Override
 	public String toString()
 	{
 		return start + "-[" + type + "]-" + (start + length);
 	}
-
+	
 	@Override
 	public void save(ByteArrayOutputStream baos) throws IOException
 	{
-		Bits.putShort(baos, (short) length);
+		Bits.putInt(baos, length);
 		baos.write(type);
 	}
 }
